refactor: extract local platform code
