+++ lib/launch-local.js
@@ -6,8 +6,6 @@ const divideWork = require('./dist');
 const { SSMS } = require('../core/lib/ssms');
 const { loadPlugins, loadPluginsConfig } = require('./load-plugins');
 
-const { ArtilleryWorker } = require('./artillery-worker-local');
-
 const EventEmitter = require('eventemitter3');
 const debug = require('debug')('core');
 
@@ -15,20 +13,8 @@ const os = require('os');
 const p = require('util').promisify;
 const _ = require('lodash');
 
-const core = require('./dispatcher');
-const { handleScriptHook, prepareScript, loadProcessor } = core.runnerFuncs;
-
-const temp = require('temp').track();
-const fs = require('fs');
-const { spawnSync } = require('child_process');
-
-const { randomUUID } = require('crypto');
-
-const sleep = require('./util/sleep');
-const path = require('path');
-
-const archiver = require('archiver');
-const AWS = require('aws-sdk');
+const PlatformLocal = require('./platform/local');
+const PlatformLambda = require('./platform/aws-lambda');
 
 async function createLauncher(script, payload, opts, launcherOpts) {
   launcherOpts = launcherOpts || {
@@ -37,432 +23,6 @@ async function createLauncher(script, payload, opts, launcherOpts) {
   };
   return new Launcher(script, payload, opts, launcherOpts);
 }
-
-class PlatformLocal {
-  constructor(script, payload, opts, platformOpts) {
-    // We need these to run before/after hooks:
-    this.script = script;
-    this.payload = payload;
-    this.opts = opts;
-    this.events = new EventEmitter(); // send worker events such as workerError, etc
-
-    this.workers = {};
-
-    return this;
-  }
-
-  async init() {
-    // 'before' hook is executed in the main thread,
-    // its context is then passed to the workers
-    const contextVars = await this.runHook('before');
-    this.contextVars = contextVars; // TODO: Rename to something more descriptive
-
-    return contextVars;
-  }
-
-  async createWorker() {
-    const worker = new ArtilleryWorker();
-
-    await worker.init();
-
-    const workerId = worker.workerId;
-    worker.events.on('workerError', (message) => {
-      this.events.emit('workerError', workerId, message);
-    });
-    worker.events.on('log', (message) => {
-      this.events.emit('log', workerId, message);
-    });
-    worker.events.on('phaseStarted', (message) => {
-      this.events.emit('phaseStarted', workerId, message);
-    });
-    worker.events.on('phaseCompleted', (message) => {
-      this.events.emit('phaseCompleted', workerId, message);
-    });
-    worker.events.on('stats', (message) => {
-      this.events.emit('stats', workerId, message);
-    });
-    worker.events.on('done', (message) => {
-      this.events.emit('done', workerId, message);
-    });
-    worker.events.on('readyWaiting', (message) => {
-      this.events.emit('readyWaiting', workerId, message);
-    });
-    worker.events.on('setSuggestedExitCode', (message) => {
-      this.events.emit('setSuggestedExitCode', workerId, message);
-    });
-
-    this.workers[worker.workerId] = {
-      proc: worker,
-      state: worker.state, // TODO: replace with getState() use
-    };
-
-    return worker;
-  }
-
-  async prepareWorker(workerId, opts) {
-    return this.workers[workerId].proc.prepare(opts);
-  }
-
-  async runWorker(workerId, contextVarsString) { // TODO: this will become opts
-    debug('runWorker', workerId);
-    return this.workers[workerId].proc.run(contextVarsString);
-
-  }
-  async stopWorker(workerId) {
-    return this.workers[workerId].proc.stop();
-  }
-
-  async getWorkerState(workerId) {
-
-  }
-
-  async shutdown() {
-    // 'after' hook is executed in the main thread, after all workers
-    // are done
-    await this.runHook('after', this.contextVars);
-  }
-
-  // ********
-
-  async runHook(hook, initialContextVars) {
-    if (!this.script[hook]) {
-      return {};
-    }
-
-    const runnableScript = loadProcessor(
-      prepareScript(this.script, _.cloneDeep(this.payload)),
-      this.opts
-    );
-
-    const contextVars = await handleScriptHook(
-      hook,
-      runnableScript,
-      this.events,
-      initialContextVars
-    );
-
-    debug(`hook ${hook} context vars`, contextVars);
-
-    return contextVars;
-  }
-}
-
-class PlatformLambda {
-  constructor(script, payload, opts, platformOpts) {
-    this.workers = {};
-    this.script = script;
-    this.payload = payload;
-    this.opts = opts;
-
-    this.events = new EventEmitter();
-
-    this.region = platformOpts.region;
-    this.testRunId = platformOpts.testRunId || randomUUID();
-    this.lambdaRoleArn = platformOpts.lambdaRoleArn;
-
-    this.artilleryArgs = [];
-  }
-
-  async init() {
-    artillery.log('Creating Lambda function');
-    const dirname = temp.mkdirSync(); // TODO: May want a way to override this by the user
-    const zipfile = temp.path({ suffix: 'zip' });
-
-    debug({dirname, zipfile});
-
-    fs.copyFileSync(path.resolve(__dirname, 'platform', 'aws-lambda', 'lambda-handler', 'index.js'), path.join(dirname, 'index.js'));
-    fs.copyFileSync(path.resolve(__dirname, 'platform', 'aws-lambda', 'lambda-handler', 'package.json'), path.join(dirname, 'package.json'));
-    fs.copyFileSync(this.opts.absoluteScriptPath, path.join(dirname, path.basename(this.opts.absoluteScriptPath)));
-    this.artilleryArgs.push('run');
-    this.artilleryArgs.push(path.basename(this.opts.absoluteScriptPath));
-    // TODO: Copy script in there. Construct Artillery args.
-
-    artillery.log('....Installing dependencies')
-    const { stdout, stderr, status, error } = spawnSync('npm install', {cwd: dirname});
-    artillery.log(stdout);
-    artillery.log('....Creating zip package');
-    await this.createZip(dirname, zipfile);
-
-    artillery.log('Creating cloud resources')
-    const bucketName = await this.ensureS3BucketExists();
-    this.bucketName = bucketName;
-
-    const s3path = await this.uploadLambdaZip(bucketName, zipfile);
-    this.lambdaZipPath = s3path;
-
-    const sqsQueueUrl = await this.createSQSQueue(this.region);
-    this.sqsQueueUrl = sqsQueueUrl;
-
-
-    if(typeof this.lambdaRoleArn === 'undefined') {
-      const lambdaRoleArn = await this.createLambdaRole();
-      this.lambdaRoleArn = lambdaRoleArn;
-    }
-
-    this.functionName = `artilleryio-${this.testRunId}`;
-    await this.createLambda({
-      bucketName: this.bucketName,
-      functionName: this.functionName,
-      zipPath: this.lambdaZipPath,
-    });
-
-    // Start SQS watcher here, emit events
-
-    debug({ bucketName, s3path, arn, sqsQueueUrl });
-  }
-
-  async createWorker() {
-    const workerId = randomUUID();
-
-    return workerId;
-  }
-
-  async prepareWorker(workerId) {
-  }
-
-  async runWorker(workerId) {
-    const lambda = new AWS.Lambda({ apiVersion: '2015-03-31', region: this.region });
-    const event = {
-      SQS_QUEUE_URL: this.sqsQueueUrl,
-      SQS_REGION: this.region,
-      WORKER_ID: workerId,
-      ARTILLERY_ARGS: this.artilleryArgs,
-    };
-    const args = Buffer.from(event);
-
-    await lambda.invokeAsync({
-      FunctionName: this.functionName,
-      InvokeArgs: args,
-    }).promise();
-  }
-
-  async stopWorker(workerId) {
-    // TODO: Send message to that worker and have it exit early
-  }
-
-  async shutdown() {
-    const s3 = new AWS.S3();
-    const sqs = new AWS.SQS({ region: this.region });
-
-    try {
-      await s3.deletObject({
-        Bucket: this.bucketName,
-        Key: this.lambdaZipPath,
-      }).promise();
-
-      await sqs.deleteQueue({
-        QueueUrl: this.sqsQueueUrl,
-      }).promise();
-    } catch (err) {
-      artillery.log(err);
-    }
-  }
-
-  // ...........
-  async createZip(src, out) {
-    const archive = archiver('zip', { zlib: { level: 9 }});
-    const stream = fs.createWriteStream(out);
-
-    return new Promise((resolve, reject) => {
-      archive
-        .directory(src, false)
-        .on('error', err => reject(err))
-        .pipe(stream)
-      ;
-
-      stream.on('close', () => resolve());
-      archive.finalize();
-    });
-  }
-
-  // TODO: reusable platform util
-  async ensureS3BucketExists() {
-    const accountId = await this.getAccountId();
-    const bucketName = `artilleryio-test-data-${accountId}`;
-    const s3 = new AWS.S3();
-
-    try {
-      await s3.listObjectsV2({Bucket: bucketName, MaxKeys: 1}).promise();
-    } catch (s3Err) {
-      if (s3Err.code === 'NoSuchBucket') {
-        const res = await s3.createBucket({ Bucket: bucketName }).promise();
-      } else {
-        throw s3Err;
-      }
-    }
-
-    return bucketName;
-  }
-
-  // TODO: reusable platform util
-  async getAccountId() {
-    let stsOpts = {};
-    if(process.env.ARTILLERY_STS_OPTS) {
-      stsOpts = Object.assign(stsOpts, JSON.parse(process.env.ARTILLERY_STS_OPTS));
-    }
-
-    const sts = new AWS.STS(stsOpts);
-    const awsAccountId = (await sts.getCallerIdentity({}).promise()).Account;
-    return awsAccountId;
-  }
-
-
-  // TODO: Add timestamp to SQS queue name. Then can GC automatically later - anything older than 24 hours.
-  async createSQSQueue() {
-    const sqs = new AWS.SQS({
-      region: this.region,
-    });
-
-    const SQS_QUEUES_NAME_PREFIX = 'artilleryio_test_metrics';
-
-    // 36 is length of a UUUI v4 string
-    const queueName = `${SQS_QUEUES_NAME_PREFIX}_${this.testRunId.slice(0, 36)}.fifo`;
-    const params = {
-      QueueName: queueName,
-      Attributes: {
-        FifoQueue: 'true',
-        ContentBasedDeduplication: 'false',
-        MessageRetentionPeriod: '1800',
-        VisibilityTimeout: '600' // 10 minutes
-      }
-    };
-
-    let sqsQueueUrl;
-    try {
-      const result = await sqs.createQueue(params).promise();
-      sqsQueueUrl = result.QueueUrl;
-    } catch (err) {
-      throw err;
-    }
-
-    // Wait for the queue to be available:
-    let waited = 0;
-    let ok = false;
-    while (waited < 120 * 1000) {
-      try {
-        const results = await sqs.listQueues({ QueueNamePrefix: queueName }).promise();
-        if (results.QueueUrls && results.QueueUrls.length === 1) {
-          debug('SQS queue created:', queueName);
-          ok = true;
-          break;
-        } else {
-          await sleep(10 * 1000);
-          waited += 10 * 1000;
-        }
-      } catch (err) {
-        await sleep(10 * 1000);
-        waited += 10 * 1000;
-      }
-    }
-
-    if(!ok) {
-      debug('Time out waiting for SQS queue:', queueName);
-      throw new Error(`SQS queue could not be created`);
-    }
-
-    return sqsQueueUrl;
-  }
-
-  async createLambdaRole() {
-    const ROLE_NAME = 'artilleryio-default-lambda-role';
-
-    const iam = new AWS.IAM();
-
-    try {
-      const res = await iam.getRole({RoleName: ROLE_NAME}).promise();
-      return res.Role.Arn
-    } catch (err) {
-      debug(err);
-    }
-
-    const res = await iam.createRole({
-      AssumeRolePolicyDocument: `{
-        "Version": "2012-10-17",
-        "Statement": [
-          {
-            "Effect": "Allow",
-            "Principal": {
-              "Service": "lambda.amazonaws.com"
-            },
-            "Action": "sts:AssumeRole"
-          }
-        ]
-      }`,
-      Path: '/',
-      RoleName: ROLE_NAME,
-    }).promise();
-
-    const lambdaRoleArn = res.Role.Arn;
-
-    await iam.attachRolePolicy({
-      PolicyArn: 'arn:aws:iam::aws:policy/service-role/AWSLambdaBasicExecutionRole',
-      RoleName: ROLE_NAME,
-    }).promise();
-
-    return lambdaRoleArn;
-  }
-
-  async createLambda(opts) {
-    const { bucketName, functionName, zipPath } = opts;
-
-    const lambda = new AWS.Lambda({ apiVersion: '2015-03-31', region: this.region });
-    const res = await lambda.createFunction({
-      Code: {
-        S3Bucket: bucketName,
-        S3Key: zipPath,
-      },
-      FunctionName: functionName,
-      Description: 'Artillery.io test',
-      Handler: 'index.handler',
-      MemorySize: 1024,
-      PackageType: 'Zip',
-      Runtime: 'nodejs16.x',
-      Timeout: 900,
-      // VpcConfig: {
-      //   SecurityGroupIds: [],
-      //   SubnetIds: [],
-      // },
-      Role: this.lambdaRoleArn,
-    }).promise();
-  }
-
-  async uploadLambdaZip(bucketName, zipfile) {
-    const key = `lambda/${randomUUID}.zip`;
-    // TODO: upload to a prefix
-    // Set lifecycle policy on that prefix - delete after 24 hours
-    const s3 = new AWS.S3();
-    const s3res = await s3.putObject({
-      Body: fs.createReadStream(zipfile),
-      Bucket: bucketName,
-      Key: key,
-    }).promise();
-
-    return key;
-  }
-}
-
-class LambdaWorker {
-  constructor() {
-
-  }
-
-  async init() {
-  }
-
-  async prepare() {
-    // nothing to do
-  }
-
-  async run() {
-    // spawn a lambda - but we don't know what it is. ie this needs to be done by Platform
-    // all of the operations here need to be done through Platform methods
-  }
-
-  async stop() {
-
-  }
-}
-
 class Launcher {
   constructor(script, payload, opts, launcherOpts) {
     this.script = script;
@@ -472,13 +32,18 @@ class Launcher {
     this.workers = {};
     this.workerMessageBuffer = [];
 
-    this.metricsByPeriod = {};
+    this.metricsByPeriod = {}; // individual intermediates by worker
+    this.mergedPeriodMetrics = []; // merged intermediates for a period
     this.finalReportsByWorker = {};
 
     this.events = new EventEmitter();
 
     this.pluginEvents = new EventEmitter();
     this.pluginEventsLegacy = new EventEmitter();
+
+    this.launcherOpts = launcherOpts;
+
+    this.periodsReportedFor = [];
 
     if (launcherOpts.platform === 'local') {
       this.count = this.opts.count || Math.max(1, os.cpus().length - 1);
@@ -494,12 +59,14 @@ class Launcher {
     if (launcherOpts.mode === 'distribute') {
       this.workerScripts = divideWork(this.script, this.count);
       this.count = this.workerScripts.length;
+    } else {
+
+      this.count = this.launcherOpts.count;
+      this.workerScripts = new Array(this.count).fill().map(_ => this.script);
     }
 
     this.phaseStartedEventsSeen = {};
     this.phaseCompletedEventsSeen = {};
-
-    // this.eventBus = new EventEmitter();
 
     this.eventsByWorker = {};
 
@@ -515,8 +82,6 @@ class Launcher {
         global.artillery.log(`[${id}]: ${error.message}`, level);
       }
     });
-
-    // TODO: We might want to expose workerPhaseStarted/Completed events
 
     workerEvents.on('phaseStarted', (workerId, message) => {
       // Note - we send only the first event for a phase, not all of them
@@ -554,9 +119,9 @@ class Launcher {
     });
 
     workerEvents.on('done', async (workerId, message) => {
-      this.workers[message.id].state = 'exited'; // TODO:
+      this.workers[workerId].state = 'exited'; // TODO:
 
-      this.finalReportsByWorker[message.id] = SSMS.deserializeMetrics(
+      this.finalReportsByWorker[workerId] = SSMS.deserializeMetrics(
         message.report
       );
 
@@ -743,20 +308,17 @@ class Launcher {
 
     setInterval(async () => {
       this.flushIntermediateMetrics();
-    }, 900).unref();
+    }, 5 * 1000).unref();
 
 
     const contextVars = await this.platform.init();
 
+    // TODO: only makes sense for "distribute" / "local"
     for (const script of this.workerScripts) {
-      const w1 = await this.platform.createWorker(); // new ArtilleryWorker();
-
-      // await w1.init();
+      const w1 = await this.platform.createWorker();
 
       this.workers[w1.workerId] = {
         id: w1.workerId,
-        // proc: w1,
-        // state: w1.state,
         script
       };
       debug(`worker init ok: ${w1.workerId}`);
@@ -773,26 +335,16 @@ class Launcher {
             payload: this.payload,
             options: this.opts
           });
-      // prepareAll.push(
-      //   w.proc.prepare({
-      //     script: w.script,
-      //     payload: this.payload,
-      //     options: this.opts
-      //   })
-      // );
     }
-    //await Promise.all(prepareAll);
     debug('workers prepared');
 
     // the initial context is stringified and copied to the workers
     const contextVarsString = JSON.stringify(contextVars);
 
+    artillery.log('Running scenarios...');
     for (const [workerId, w] of Object.entries(this.workers)) {
       await this.platform.runWorker(workerId, contextVarsString);
-      //runAll.push(w.proc.run(contextVarsString));
     }
-
-    //await Promise.all(runAll);
 
     debug('workers running');
   }
