+++ lib/platform/local/artillery-worker-local.js
@@ -6,7 +6,7 @@ const EventEmitter = require('eventemitter3');
 const { Worker } = require('worker_threads');
 const path = require('path');
 
-const awaitOnEE = require('./util/await-on-ee');
+const awaitOnEE = require('../../util/await-on-ee');
 
 const STATES = {
   initializing: 1,
