+++ lib/platform/local/worker.js
@@ -16,14 +16,14 @@ const {
   threadId
 } = require('worker_threads');
 
-const core = require('./dispatcher');
+const core = require('../../dispatcher');
 const createRunner = core.runner;
 const debug = require('debug')('artillery:worker');
 
 const path = require('path');
 
-const { SSMS } = require('../core/lib/ssms');
-const { loadPlugins, loadPluginsConfig } = require('./load-plugins');
+const { SSMS } = require('../../../core/lib/ssms');
+const { loadPlugins, loadPluginsConfig } = require('../../load-plugins');
 
 const EventEmitter = require('eventemitter3');
 const p = require('util').promisify;
